requirements.txt
numpy>=1.21.0
pandas>=1.3.0

torch>=1.9.0
torchvision>=0.10.0
torchtext>=0.10.0
scikit-learn>=0.24.2

biopython>=1.79
rdkit>=2021.09.4

matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3

# Optional dependencies for GPU support
cudatoolkit>=11.1  # Match your PyTorch installation

# Scheduling and optimization
optuna>=2.10.0

